+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## 2.4.16.0
+
+### Added
+
+- dx_get_db_env has a new column showing last VDB refresh time
+- TDE support for an Oracle MT
+- OAuth2 support - Bearer token access
+
+
+### Changed
+ - dx_get_vdbsize - documentation fix
+ - dx_ctl_env - set TDE password for CDB
+ - dx_provision_db - support for Oracle MT TDE
+ - dx_get_osversion - print upgrade verification results
+
 ## 2.4.15.1
 
 ### Added